Refactor password page
