--- packages/web/src/pages/sign-on/components/desktop/LeftContentContainer.module.css
@@ -1,4 +1,0 @@
-.root {
-  background-color: var(--harmony-white);
-  text-align: left;
-}
