+++ packages/web/src/pages/sign-up-page/SignUpPage.tsx
@@ -4,9 +4,9 @@ import { useSelector } from 'react-redux'
 import { Redirect, Route, RouteProps, Switch } from 'react-router-dom'
 
 import { getSignOn } from 'common/store/pages/signon/selectors'
-import SignOnPageState from 'common/store/pages/signon/types'
 import { useMedia } from 'hooks/useMedia'
-import { useRouteMatch } from 'hooks/useRouteMatch'
+import { NavHeader } from 'pages/sign-up-page/components/NavHeader'
+import { determineAllowedRoute } from 'pages/sign-up-page/utils'
 import { AppState } from 'store/types'
 import {
   SIGN_UP_ARTISTS_PAGE,
@@ -15,14 +15,11 @@ import {
   SIGN_UP_GENRES_PAGE,
   SIGN_UP_HANDLE_PAGE,
   SIGN_UP_PAGE,
-  SIGN_UP_PASSWORD_PAGE,
-  SignUpPath,
-  TRENDING_PAGE
+  SIGN_UP_PASSWORD_PAGE
 } from 'utils/route'
 
-import { MobileNavHeader } from './components/MobileNavHeader'
 import { ProgressHeader } from './components/ProgressHeader'
-import { CreateEmailPage } from './pages/CreateEmailPage/CreateEmailPage'
+import { CreateEmailPage } from './pages/CreateEmailPage'
 import { CreatePasswordPage } from './pages/CreatePasswordPage'
 import { FinishProfilePage } from './pages/FinishProfilePage'
 import { PickHandlePage } from './pages/PickHandlePage'
@@ -32,84 +29,6 @@ import { SelectGenrePage } from './pages/SelectGenrePage'
 const messages = {
   metaTitle: 'Sign Up',
   metaDescription: 'Create an account on Audius'
-}
-
-/**
- * Checks against existing sign up redux state,
- * then determines if the requested path should be allowed or not
- * if not allowed, also returns furthest step possible based on existing state
- */
-const determineAllowedRoute = (
-  signUpState: SignOnPageState,
-  requestedRoute: string | SignUpPath // this string should have already trimmed out /signup/
-): {
-  allowedRoutes: string[]
-  isAllowedRoute: boolean
-  correctedRoute: string
-} => {
-  const attemptedPath = requestedRoute.replace('/signup/', '')
-  // Have to type as string[] to avoid too narrow of a type for comparing against
-  let allowedRoutes: string[] = [SignUpPath.createEmail] // create email is available by default
-  if (signUpState.email.value) {
-    // Already have email
-    allowedRoutes.push(SignUpPath.createPassword)
-  }
-  if (signUpState.password.value || signUpState.useMetaMask) {
-    // Already have password
-    allowedRoutes.push(SignUpPath.pickHandle)
-  }
-  if (signUpState.handle.value) {
-    // Already have handle
-    allowedRoutes.push(SignUpPath.finishProfile)
-  }
-  if (signUpState.name.value) {
-    // Already have display name
-    // At this point the account is fully created & logged in; now user can't back to account creation steps
-    allowedRoutes = [SignUpPath.selectGenres]
-  }
-
-  // TODO: These checks below here may need to fall under a different route umbrella separate from sign up
-  if (signUpState.genres) {
-    // Already have genres selected
-    allowedRoutes.push(SignUpPath.selectArtists)
-  }
-
-  if (signUpState.followArtists?.selectedUserIds?.length >= 3) {
-    // Already have 3 artists followed
-    // Done with sign up if at this point so we return early (none of these routes are allowed anymore)
-    return {
-      allowedRoutes: [],
-      isAllowedRoute: false,
-      correctedRoute: TRENDING_PAGE
-    }
-  }
-
-  const isAllowedRoute = allowedRoutes.includes(attemptedPath)
-  // If requested route is allowed return that, otherwise return the last step in the route stack
-  const correctedPath = isAllowedRoute
-    ? attemptedPath
-    : allowedRoutes[allowedRoutes.length - 1]
-
-  return {
-    allowedRoutes,
-    isAllowedRoute,
-    correctedRoute: `/signup/${correctedPath}`
-  }
-}
-
-const useIsBackAllowed = () => {
-  const match = useRouteMatch<{ currentPath: string }>('/signup/:currentPath')
-  const existingSignUpState = useSelector((state: AppState) => getSignOn(state))
-  if (match?.currentPath) {
-    const { allowedRoutes } = determineAllowedRoute(
-      existingSignUpState,
-      match?.currentPath
-    )
-    const currentRouteIndex = allowedRoutes.indexOf(match.currentPath)
-    const isBackAllowed = allowedRoutes.length > 1 && currentRouteIndex > 0
-    return isBackAllowed
-  }
-  return false
 }
 
 /**
@@ -138,7 +57,6 @@ export function SignUpRoute({ children, ...rest }: RouteProps) {
 
 export const SignUpPage = () => {
   const { isDesktop } = useMedia()
-  const isBackAllowed = useIsBackAllowed()
 
   return (
     <>
@@ -146,7 +64,7 @@ export const SignUpPage = () => {
         <title>{messages.metaTitle}</title>
         <meta name='description' content={messages.metaDescription} />
       </Helmet>
-
+      <NavHeader />
       <Switch>
         <Route exact path={SIGN_UP_PAGE}>
           <Redirect to={SIGN_UP_EMAIL_PAGE} />
@@ -166,33 +84,27 @@ export const SignUpPage = () => {
             SIGN_UP_ARTISTS_PAGE
           ]}
         >
-          <Paper direction='column' w='100%' h={864}>
-            {isDesktop ? (
-              <ProgressHeader />
-            ) : (
-              <MobileNavHeader isBackAllowed={isBackAllowed} />
-            )}
-            <Switch>
-              <SignUpRoute exact path={SIGN_UP_HANDLE_PAGE}>
-                <PickHandlePage />
-              </SignUpRoute>
-            </Switch>
-            <Switch>
-              <SignUpRoute exact path={SIGN_UP_FINISH_PROFILE_PAGE}>
-                <FinishProfilePage />
-              </SignUpRoute>
-            </Switch>
-            <Switch>
-              <SignUpRoute exact path={SIGN_UP_GENRES_PAGE}>
-                <SelectGenrePage />
-              </SignUpRoute>
-            </Switch>
-            <Switch>
-              <SignUpRoute exact path={SIGN_UP_ARTISTS_PAGE}>
-                <SelectArtistsPage />
-              </SignUpRoute>
-            </Switch>
-          </Paper>
+          {isDesktop ? <ProgressHeader /> : null}
+          <Switch>
+            <SignUpRoute exact path={SIGN_UP_HANDLE_PAGE}>
+              <PickHandlePage />
+            </SignUpRoute>
+          </Switch>
+          <Switch>
+            <SignUpRoute exact path={SIGN_UP_FINISH_PROFILE_PAGE}>
+              <FinishProfilePage />
+            </SignUpRoute>
+          </Switch>
+          <Switch>
+            <SignUpRoute exact path={SIGN_UP_GENRES_PAGE}>
+              <SelectGenrePage />
+            </SignUpRoute>
+          </Switch>
+          <Switch>
+            <SignUpRoute exact path={SIGN_UP_ARTISTS_PAGE}>
+              <SelectArtistsPage />
+            </SignUpRoute>
+          </Switch>
         </SignUpRoute>
       </Switch>
     </>
